# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Авто машины зогсоолын системийн API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface VehicleEnterRequest {
  /** Vehicle license plate number */
  plateNumber: string;
}

export type ParkingRecordStatus =
  (typeof ParkingRecordStatus)[keyof typeof ParkingRecordStatus];

export const ParkingRecordStatus = {
  active: "active",
  pending_payment: "pending_payment",
  paid: "paid",
  completed: "completed",
} as const;

export interface ParkingRecord {
  id: number;
  plateNumber: string;
  entryTime: string;
  exitTime?: string | null;
  fee?: number | null;
  status: ParkingRecordStatus;
  durationMinutes?: number | null;
}

export interface ParkingRecordsResponse {
  records: ParkingRecord[];
  total: number;
}

export interface ExitResponse {
  record: ParkingRecord;
  fee: number;
  durationMinutes: number;
  /** QR code data URL for payment */
  qrCode: string;
}

export interface QRResponse {
  /** QR code data URL */
  qrCode: string;
  fee: number;
  recordId: number;
}

export interface PaymentResponse {
  success: boolean;
  record: ParkingRecord;
  gateOpened: boolean;
  message: string;
}

export interface ParkingStatusResponse {
  totalSpaces: number;
  occupiedSpaces: number;
  availableSpaces: number;
  occupancyRate: number;
}

export interface RevenueByDay {
  date: string;
  revenue: number;
  vehicles: number;
}

export interface AdminStatsResponse {
  totalRevenue: number;
  totalVehicles: number;
  activeVehicles: number;
  averageDurationMinutes: number;
  revenueByDay: RevenueByDay[];
}

export interface TariffResponse {
  ratePerHour: number;
  minimumFee: number;
  freeMinutes: number;
}

export interface UpdateTariffRequest {
  ratePerHour: number;
  minimumFee: number;
  freeMinutes: number;
}

export interface CapacityResponse {
  totalSpaces: number;
}

export interface UpdateCapacityRequest {
  totalSpaces: number;
}

export type GetParkingRecordsParams = {
  /**
   * Filter by status
   */
  status?: GetParkingRecordsStatus;
  limit?: number;
  offset?: number;
};

export type GetParkingRecordsStatus =
  (typeof GetParkingRecordsStatus)[keyof typeof GetParkingRecordsStatus];

export const GetParkingRecordsStatus = {
  active: "active",
  completed: "completed",
  all: "all",
} as const;

export type GetAdminStatsParams = {
  /**
   * Start date (ISO string)
   */
  from?: string;
  /**
   * End date (ISO string)
   */
  to?: string;
};