# lib/api-zod/src/generated/types/adminStatsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Авто машины зогсоолын системийн API
 * OpenAPI spec version: 0.1.0
 */
import type { RevenueByDay } from "./revenueByDay";

export interface AdminStatsResponse {
  totalRevenue: number;
  totalVehicles: number;
  activeVehicles: number;
  averageDurationMinutes: number;
  revenueByDay: RevenueByDay[];
}