# lib/api-zod/src/generated/types/exitResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Авто машины зогсоолын системийн API
 * OpenAPI spec version: 0.1.0
 */
import type { ParkingRecord } from "./parkingRecord";

export interface ExitResponse {
  record: ParkingRecord;
  fee: number;
  durationMinutes: number;
  /** QR code data URL for payment */
  qrCode: string;
}