# lib/api-zod/src/generated/types/parkingRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Авто машины зогсоолын системийн API
 * OpenAPI spec version: 0.1.0
 */
import type { ParkingRecordStatus } from "./parkingRecordStatus";

export interface ParkingRecord {
  id: number;
  plateNumber: string;
  entryTime: Date;
  exitTime?: Date | null;
  fee?: number | null;
  status: ParkingRecordStatus;
  durationMinutes?: number | null;
}