# lib/api-zod/src/generated/types/parkingRecordsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Авто машины зогсоолын системийн API
 * OpenAPI spec version: 0.1.0
 */
import type { ParkingRecord } from "./parkingRecord";

export interface ParkingRecordsResponse {
  records: ParkingRecord[];
  total: number;
}