# lib/api-zod/src/generated/types/paymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Авто машины зогсоолын системийн API
 * OpenAPI spec version: 0.1.0
 */
import type { ParkingRecord } from "./parkingRecord";

export interface PaymentResponse {
  success: boolean;
  record: ParkingRecord;
  gateOpened: boolean;
  message: string;
}