# lib/api-zod/src/generated/types/tariffResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Авто машины зогсоолын системийн API
 * OpenAPI spec version: 0.1.0
 */

export interface TariffResponse {
  ratePerHour: number;
  minimumFee: number;
  freeMinutes: number;
}